Reorganize file locations
