+++ TelltaleEditor/Source/ToolLibrary/Source/Core/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_subdirectory(Platform/${TTE_TARGET_PLATFORM})
