+++ TelltaleEditor/Source/ToolLibrary/Source/Core/Platform/Win64/Thread.cpp
