+++ TelltaleEditor/Source/ToolLibrary/Source/Scheduler/CMakeLists.txt
@@ -1,2 +1,1 @@
 target_sources(${TARGET_NAME} PRIVATE JobScheduler.cpp)
-add_subdirectory(Platform/${TTE_TARGET_PLATFORM})
